Play across exercises more accurately
A few thoughts

1. We still probably need a transition screen?
2. The metronome may need more events to really make things seamless
