+++ src/core/practice-session.ts
@@ -1,5 +1,10 @@
 import { BPM, ZeroToOneHundred } from "@/core/base";
 
+import { notes } from "@/core/notes";
+import { shuffle } from "remeda";
+
+type NotePool = { kind: "circle-of-fourths" };
+
 /**
  * This single exercise will cover all of the existing exercises.
  */
@@ -7,7 +12,7 @@ export type ExerciseTimedPitchesConfig = {
   title: string;
   description: string;
   tempo: BPM;
-  notes: string[];
+  notes: NotePool;
   beatsPerNote: number;
   numberOfCycles: number;
   metronomeVolume: ZeroToOneHundred;
@@ -37,5 +42,48 @@ export function makePracticeSession(
     description: "",
     exercises: [],
     ...overrides,
+  };
+}
+
+/**
+ * A generator which produces "random" notes, but will not repeat a note until
+ * all notes have been used.
+ * After all notes are used once, the pool of notes automatically resets to the initial set.
+ *
+ * NOTE: Currently unused, but keeping for the near future...
+ */
+export function makeNoteGenerator(initialNotes: string[] = [...notes]) {
+  let availableNotes: string[] = [];
+
+  return {
+    nextNote() {
+      if (availableNotes.length === 0) {
+        availableNotes = shuffle(initialNotes);
+      }
+
+      const [note, ...nextAvailable] = availableNotes;
+
+      availableNotes = nextAvailable;
+      return note;
+    },
+  };
+}
+
+export function getDefaultTimedPitchesExercise(
+  overrides: Partial<ExerciseTimedPitchesConfig> = {},
+): ExerciseTimedPitches {
+  return {
+    type: "timed-pitches",
+    config: {
+      title: "",
+      description: "",
+      tempo: 80,
+      beatsPerNote: 12,
+      numberOfCycles: 1,
+      notes: { kind: "circle-of-fourths" },
+      metronomeVolume: 50,
+      pitchVolume: 50,
+      ...overrides,
+    },
   };
 }
